+++ composer.json
@@ -19,7 +19,8 @@
   "autoload": {
     "psr-4": {
       "App\\": "src/"
-    }
+    },
+    "files": ["src/dump.php"]
   },
   "autoload-dev": {
     "psr-4": {
