--- src/Conexao.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-use \PDO;
-
-class Conexao {
-  public static function criar(): PDO {
-    return require('db_connect.php');
-  }
-}
