+++ src/login_admin/index.php
@@ -1,7 +1,7 @@
 <?
 use App\Pagina;
 
-require_once('../../vendor/autoload.php');
+require_once(__DIR__.'/../../vendor/autoload.php');
 
 $pag = new Pagina();
 ?>
